requirements.txt
numpy
cffi==1.14.2
https://github.com/telejesus2/PyRep/archive/master.zip
numpy
Pillow
pyquaternion
html-testRunner
setuptools
natsort
https://github.com/telejesus2/RLBench/archive/master.zip
torch==1.7.1
torchaudio==0.7.2
torchvision==0.8.2
termcolor
PyYAML
tensorboard
scipy
psutil
seaborn
